refactor: `Parseable.store_ingestor_metadata()`
